App divided in modules
Lazy loading with preloadedStrategy implemented

+++ courseProjectOne/src/app/recipe-book/recipe-detail/recipe-detail.component.ts
@@ -46,8 +46,10 @@ export class RecipeDetailComponent implements OnInit {
   }
 
   onDelete() {
-    this.recipeService.deleteRecipe(this.id)
-    this.router.navigate(['/recipes'])
+    if (confirm('Are you sure you want to Delete?')) {
+      this.recipeService.deleteRecipe(this.id)
+      this.router.navigate(['/recipes'])
+    }
   }
 
 }
